remove feature flag
